requirements.txt
dotenv==0.9.9
PyJWT==2.10.1
python-dotenv==1.2.1
